pyproject.toml
[project]
name = "voice-ai-cli"
version = "0.1.0"
description = "Voice AI CLI agent that makes outbound phone calls and has casual conversations"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents~=1.4",
    "livekit-api",
    "livekit-plugins-deepgram",
    "livekit-plugins-openai",
    "livekit-plugins-cartesia",
    "livekit-plugins-silero",
    "livekit-plugins-turn-detector",
    "livekit-plugins-noise-cancellation",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
voice-ai = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
